Fixes scrolling error with cross-bar tied notes; some logging changes
Error had to do with building the animation object that is eventually
passed to the SSScrollView. This animation directs the speed of scrolling,
and the offset into the ContentView at any time - the sync between the
audio playback of the score and the visual scrolling.

This animation is built by setting the x-offset into the scrolled-view
at specific times. The code uses the note's "startTImeInMeasure" for this
calculation.

In SeeScore's rep of the XML, the first note (of the pair) contains the
duration of the entire pair, so it's sufficient to include only the first
note's data.

The problem was the inclusion of the data from the second note of a
cross-bar tied note pair. Its "startTImeInMeasure" was negative - since it
started in the previous measure, so this was used for the "time" portion of
the animation entry, but the X-offset of the 2nd note was used for the
"position" portion.

With two entries at the same time, the second one won, and scrolling
jumped to the second note's x-offset.

Fix: skip the second note.

Also added some additional, conditional logging for debugging.