First checkin of Sound and Note Performance Analysis
First stable, working code for (my) Sound and Note Performance Analysis
- creates a detailed Sound object for every sound detected, and saves these for analysis
- creates a PerfNote object for every Note in the Score (links to sound if they overlap)
- links Sounds to Notes, if begin times pass criteria and neither already linked
- detects the case where the student is playing legato and transistions to a new note. Stops current sound and creates a new one

Notes and Sounds can be inspected (to be implemented later) to see how well the student met the expectations of each note.

- also, first rudementary version of Overlay view on SSScrollView. Will be modified, and can be taken out if not desired - not linked directly to other work.

(Omitting XCode proj file for now, as had to change locally for Team and signing issues. Assume it will be checked in when actual PR is issued.)