Fixes for NoteInfo alert, partials, timing
Fixed issues with the NoteInfo alert:
- Was using wrong NoteID (Concert instead of
Transposed) in several places.
- Analysis wasn't using correct analyzer, one that
includes the acutal routine to check for partials

The transposition offset (2 for trumpet) was
not set properly in the PerfomanceAnalysisMgr
for the first calls to get it. Fixed.

I'm seeing a 40ms latency; added a 40ms
adjustment  to Sound start and end to correct for
this. (Also had to adjust the xPos for Sound Boxes
on Overlay view.

I think the metronome drawing is off (the dots). Also added a 40ms adjustment.

changed the color and transparency of metronome
normal dots (made them much lighter).

Added more testing for partials.