Fleshing out of Performance Tracking and Analysis
I’ve included an overview, PerformanceAnalysisOverview.pdf, under the folder “Documents”. I recommend reading this before diving in to the code.

There are three distinct phases, and three distinct subsystems to represent them.

PerformanceTracking
- Creates Sound objects representing the student’s performance, which include attack time, pitch, etc.
- Creates Note objects representing a note on the score, the expected performance.
- Will link a Sound and Note together if the Sound’s attack is in the window expected by the Note.
- It is then possible to grade the performed note by pitch, attack, and duration.

PerformanceAnalysis
- Each note is visited and graded by a PitchAnalyzer and RhythmAnalyzer.
- The tables used to define the zones for grading are generated at app startup, but this can be moved so that the tables can be rebuilt at lesson change. This enables the code to make tighter windows for acceptance as the student improves.

PerformanceIssue identification
- A performance criteria is chosen (what are we concentrating on?): Pitch, Attack, Duration, or Individual (the worst offense for each note).
- An issue is created for each note, using the criteria
- The issues are sorted by severity, worst is the first entry
- Each issue is assigned a Video ID correlating to the error.

When the PerformanceIssue phase is done, the app takes the first entry in the Issues array, instructs the SeeScore scrolling view to scroll to the note and highlight it, then it launches the VideoHelpView which plays the video.

Below is the hierarchy of the subsystems and the files that support them:

PerfomanceTrackingMgr.swift
 - NoteFreqRangeTable.swift
 - NotePitchAnalysysCriteria.swift
 - PerformanceSound.swift
 - PerformanceNotes.swift
 - TrumpetPartials.swift

PerfomanceAnalysisMgr.swift
 - NotePerformanceAnalyzer.swift
   - NotePtichPerformanceAnalyzer.swift
   - NoteRhythmPerformanceAnalyzer.swift

PerformanceIssueMgr.swift
 - PerfAnalysisSettingsPopup.swift
 - VideoHelpResources.swift
 - VideoHelpView.swift

Common to all:
NoteIDs.Swift
PerformanceAndTrackingDefs.swift
   - contains refs and structs used by the subsystems. Placed here in one
     file for ease of access for tuning.

There is a button, in the top left corner of TuneExerciseVeiwController.swift that will display the PerfAnalysisSettingsPopup. To turn on/off the display of this button, set kMKDebugOpt_ShowDebugSettingsBtn in PerformanceAndTrackingDefs.swift accordingly.

One last comment: I’ve been reluctant to remove the existing analysis code from TuneExerciseVeiwController.swift, including the comments left by previous devs. The approach was different, the data at the end of a performance is different, and there may be some code in there we want to keep.

My own code in other files is much cleaner as well.

This combo makes TuneExerciseVeiwController.swift quite bloated.  Hope I can remedy this soon, or someone else.