Fixing build issues after merge and xcode/swift upgrade
Numerous changes to the Xcode project, the Carthage Cartfile, and Swift files.

Retrieiving and building SwiftyJSON is now managed by Carthage, but the project is itself also embedded in the XCode project, but without a build step to actually build it. In addition to embedding the framework (built by Carthage), the project was also trying to include the framework from the location where it would have been if built within our project.

I removed the erroneous include path from all configs, and added it to the "built-by-Carthage" frameworks. (AudioKit is in this group; it is managed and built by Carthage, so this is consistant.)

(I plan to remove the SwiftyJSON project from the FirstStage project, having Jay's agreement, in a future commit. Keeping this change simple for now.)

Other Xcode project file changes:
- Removed a few file double includes
- In Release config, removed a json file from the "Compile" list that is not a compilable file

Cartfile:
- This file declares the projects to be managed and compiled by Carthage.  I was having compatibility issues between the current versions of SwiftyJSON and AudioKit and my version of Xcode/Swift. So I "pinned" SwiftyJSON and AudioKit to versions compatible with current environment. Other devs will need to upgrade this in the future to match their current environment. 

*.swift file changes:
- Several String and Font changes required when upgrading from Swift 3.2 to 4.0
- fixed several warnings
- added "@objc" prefix to several functions used selectors in target declarations