Get Longtones view working again
SeeScore is needed to display the note to play, and to play the note if the student taps "Tap to hear note". There were numerous issues with this view. I'm pretty sure this was working at one point, with an earlier version of SeeScore. But - I think - a newer version of SeeScore was brought in to address the requirements for the TuneExercise scene, and this required changes to LongTones view that the dev didn't want to make. Instead, sections were commented out.

Specifically:

LoadFile (the music XML file) method from TuneExer had to be copied, pasted, and modified.

Providing a playback synth for "tap to hear" required changes borrowed from TuneExer.

Playback was off by a whole tone. (Since trumpet is transposed by this amount.) Had to modify the two MusicXML files to transpose each note for display only.