Performance Tracking and Analysis now detects sounds played during rests
Modified PerformanceNote: 
- Created a base class, PerformanceScoreObject with most of the properties needed by derived classes
- Modified PerformanceNote to derive from PerformanceScoreObject
- Created new PerformanceRest class, derive from PerformanceScoreObject

Added new enum members for rating and scoring Rests, invoking Alerts or Videos, and added these where appropriate in Performance Analysis and Rating code. 

Added code that creates and manages PerfRests, in response to an event from SeeScore that a rest occurred in the score.

Added code to display rests, correct or bad, on the Overlay view of SeeScore.

Etc.