"Ejector Seat" - stop performance if error with note or rest > threshold
Previously, the code:
- waited until student finished (or stopped via button press) performing the tune in TuneExercise view
- and then went through all the notes scoring each
- sorted issues by severity
- and finally popped up a video if there was a problem.

That meant that the student could play the entire tune only to discover errors in the very beginning kept them from passing.

Now code scans each note/rest when it completes, and then determines right them if there is an error that will fail the performance. If so, playback is stopped and video or alert is shown immediately.