Adding dynamic control of Sound Start delay, based on HW vs Simulator
When using the simulator, there is a noticeable delay between when a sound is made and when AudioKit detects it. I think this is due to the fact that the Simulator is running in a virtual machine, and is using the Mac's mic, etc.  

I had added a delay-compensation constant, used to determine if the attack of a played note is accurate, but this was constant whether running on HW or in the Simulator.  

This change selects between two delay values based on if the app is running in the simulator vs a real device.

(The HW value is currently 0. It can be adjusted if it seems that there actually needs to be some time compensation when using HW as well.)