Getting LongTones view working; Scheduler first checkin; misc changes
Main purpose of checkin is changes to LongTones screen, with supporting files.

Changed Balloon.swift to use balloon images from MonkeyTones; added transitions for expanding and exploding ballon, etc.  

Added StarScoreView, which for the time being uses bananas instead of stars (but could easily use star images).  

Fully baked version of PerformanceScoreObjectScheduler. Other files that call it are not checked in with this commit.  (So that I can checkin the proj file without having to remove it - this was the only file in the proj file not having to do with the LongTone changes.)

Misc changes/fixes to ThresholdMgr and PerfTrackingDefs file.

Note: I was working on many things at once, not discreet changes. I'm only checking in the changes for Longtones (with the exception of the scheduler).  I'm very confident this will build, but if not, broken build issues will be resolved with subsequent commits.