checkin for build 1.0.64, 4th submission
Primarily added reminder every 3 weeks to check App Store for  new version.

Added other stuff, to check if the is a new version, but decided to take that and just do a reminder.

Added new key for SeeScore (to allow transposition), but decided not to use it, in case it screws up existing code.  Commented out.  

Also added a call to Transpose in TuneExerciseVC, but commented it out as well.