XML files (from Shawn) to complete Levels 1-10 for brass
These are the separate TUN, SLR, PTY, and PRP XML files for each instrument.

The Trombone, Horn, and Tuba files are new from Shawn. The Trumpet files are copied from the original (with full names) and renamed to the new format.

They are placed into the Trombone, Tuba, Trumpet, and French Horn folders.

When it was just trumpet, these files used to have the Tune's title as the filename (e.g., "Mexican Hat Dance.xml). (They also, in some cases, included "Trumpet" or some such as part of the filename.) 

Now the filename is the exercise name (e.g., TUN027.xml).  In this way the corresponding exercises in each instrument folder are named the same, and all that's needed is to switch the folder for the path to the filename.

Also included are the LongTone support files. These are single measure, multiple parts files. Each part is a single note. The LongtoneVC accesses this file, and displays the part corresponding to the requested note. There is a file for sharps and a file for flats.