First group of Instrument specific XML files
For extending to all brass instruments, each instrument now has a directory under "XML Tunes". 

For example:
\XMLTunes\FrenchHorn\TUN001.xml
\XMLTunes\Tuba\TUN001.xml
\XMLTunes\Trumpet\TUN001.xml
etc.

(The code supporting this directory change will be checked in in a subsequent commit. For now, just checking in the file to avoid an enormous, cluttered commit.)

These are the files needed for the first 10 levels. Checkins will follow for the remaining levels.