Upgrading to SeeScore version 330
Incorporated SeeScore version 330 into app.

The most challenging aspect of this was upgrading the sample app files that were used by the app, starting about three years ago.

Most 3rd party kits, like SeeScore, include a sample app. The sample code uses the library, making calls to the API, etc.

The existing sample code did not meet the needs of David (one of the original devs), and he changed it a lot, quite a bit in the case of SSScrollView.h/m. 

And a lot has changed in the SeeScore API and sample code in three years. So the challenge was to go through all the changes to 5 sample code file (used by the app) and try to figure out which changes to override b/c of David’s intent. In particular, the code that dynamically determines the height of the embedded scroll view. (Without correctly choosing David’s changes, a long score filled the entire screen, was only 20 pixels high, and didn’t scroll at all.)

The changes to the API required lots of changes in our code that interacts with the API as well.