Dealing with V300 SeeScore Framework files that are > 100MB
GitHub has an individual file size limit of 100MB/file. 

For previous releases of SeeScore, in particular the framework files, this wasn't an issue. We zipped up the frameworks files - the results were less than  100MB - and checked these in.

But for V330, the compressed (zipped) file for the Universal framework was 110MB, and GitHub stopped accepting it.

The solution is to put it in external storage, which is what I did. These files now live on Google Drive, 

So this checkin::
- deletes the zipped framework files from the AppStore-Release and Universal folders under SeeScore.
- (not directly part of this checkin):  Placed the frameworks files up on Google Drive    
- includes  placeholder text files, and instruction text file in parent dir.