Mostly fixing issues with SeeScore rendering (too wide, etc.)
Main thrust was to address issues encountered with new SeeScore release. Scores are now displayed with a fixed-measure width. Every note takes up the same space in a measure, and the widest measure determines the width for all measures in the score. 

First issue: 
The shortest note in the score determines this width; e.g., just one 1/8 note make the score twice as wide as a score with only 1/4 notes. A score with only whole notes is very short in comparison. (There is some margin width for each measure, but this is constant.)

Bottom line: For songs with 1/8 notes, the measures are quite wide. This exacerbates the problem below, but also makes it hard to see more than two measures (or less!) of the score at a time. At fast tempos, this is really an issue.

Second issue: 
(This may have been a problem all along, but was not seen because the first issue eposes it?) There appears to be a maximum width of 8000 pixels for a pre-rendered scrolling score. (I’m not sure if this is an iOS/Mac limitation or something within SeeScore. I wrote James and he didn’t say it was SeeScore’s limitation.

So some of the longer scores with 1/8 notes, e.g., Children’s March, won’t display.

Third issue:
The SeeScore example code that David modified has a loop that is responsible for determining the width of the fully rendered score. This is necessary to build the scrolling animation, and may be necessary to display anything at all, as we are not using SeeScores’s (and Apple’s) recommended method of rendering each part of the view as it comes into view. (It actually might not be possible to do this, and SeeScore thinks in “systems”, which is normally one staff(s) of many on a fixed width page. Instead, we have the whole score display as one system, and we scroll that.

The loop takes screen’s width (as a start) and keeps adding 100 to it until the “it’s all here” flag is TRUE in the return from SeeScore’s “render score” method. For scores that can be 8000 or more pixels wide, this can take 80+ iterations, which can take 3-5 seconds.

So . . .

1) Created a way to feed values for NoteWidth and Magnification to the score building code. By tweaking these down from 30 to 25 pixels and 1.4 to 1.2 respectively, I was able to get all scores below 8000 pixels. 

Also, this improves the density of notes/measure, so that more can be seen on the screen at a given time.

2) Created a way to feed the initial starting width to the score building code. By observing the end size of the rendered view (using the debugger), (say 7200 pixels), I can feed a slightly smaller value (say 6800) in, so the code only loops 4-5 times, instead of 80; this greatly speeds up the rendering. 

3) Added each of these fields (NoteWidth, Magnification, and FrameWidth) to every entry in XMLFileMapping.json (300-500 entries?). 99% have values of “0”.

4) Modified file-mapping code to read these values from the file and supply them to the score building code.

5) The score building code looks at all of these values and:
- if they are 0, it uses the current default values for each (code acts just as it did before)
- if they are non-0, it uses this value instead of the default.

6) In this way we can set values in the json for problematic files. I made changes for 4 files that were having issues and were not displaying at all (going ver the 8000 value) or were too wide to work with. We will need to identify others that are an issue and make changes for them.

Misc:

- Bug fix: When adding a pause to the scrolling, so that the cursor remains closer to the center, I created a bug where the maximum star score possible was only 3 out of 4. Fixed it. 

- In the InternalSettings view, changed the SoundStart slider’s minimum value from 0.1 to 0.2 seconds.