Changes for 6/2020 Upgrade, before adding xcconfig
These are the changes to get the app using the versions of Xcode and iOS SDK to meet Apple's submission requirements for 6/30/2020.

Build-wise, most everything is worked out except for an issue with AEXML. For some reason instead of a combined library for debug-mode, they have mow created two versions: one for debug-ios-simulator, and one for debug-ios (an actual device).  

This checkin is the changes to date before I try to resolve this AEXML issue.