Changes made for V2.1.0 Release: ClassKit/SchoolWork integration
Changed end-of-trial period to 8/1/2020.

ClassKit/SchoolWork integration

Since this is not context-free as Apple recommends, I have to make update calls in ClassKit for all exercises at the start of a Day, and after each exercise. I’m not going to try to explain just why this is and how this has to work here - to much to explain. But here are some points:

- They could have done some of the Day on another device.
- There is no State of Things coming to me - the app os flying blind. 
- The app, by default, allows the student to accomplish some of the work before it is assigned. So I need to push it all the first time the Day is invoked under SchoolWork.
- SchoolWork won’t accept a “grade” lower than the current one. So my sending a worse score won’t override a better existing one.

publishContextsIfNeeded() 

registers ClassKit Contexts in this hierarchy: 
- - Levels
- - - - Days within the Levels

There is a call in AppDelegate, continue/restorationHandler, that takes an incoming call from SchoolWork; one of the params contains the ClassKitPath, which is the Level/Day to open, and Navigation  

There are 15-ish “ActivityItems” associated with each Day, but they are

ClassKitPath is passed down as navigation occurs through the App’s view hierarchy, but it isn’t until the Day that the context is looked up. 

One in the Day, all processing is handled by  is all handled by the  (new) DayCKActivityManager, which creates 15-or-so “ActivityItems” associated with the Day (e.g, average Star Score, total percent done, Avg Star Score for Tunes, LongTones, etc.). DayCKActivityManager is responsible for making calls to update these values in the SchoolWork environment.  

DayCKActivityManager calls the (new) DayScoreSummarizer, which gathers the scores and status fo each exercise in the Day, divides them into categories (Tunes, LongTones, Scales, Rhythm Party, Rhythm Prep, etc.), and calculates average star scores and percent done for each category. 

The Score File data structs were changed for this release, adding fields needed for ClassKit - this new config is Version 4. This required:
- new methods for setting/getting these new fields
- adding code that updates an on-disk V2/V3 Score file to a V4 score file.



Bug Fix: Shawn noticed duplicate TUN006 entries in Days 2 and 4 of Levels 2, 3, and 4. We changed the exercise lists for those days to fix this.


One thing I broke:  In making changes necessary for the navigation from within AppDelegate::continue/restorationHandler() to work, The first Onboarding slides no longer work. The App goes straight to the Main Screen when first opened.

I did try to fix this, but we were down to the wire, and I was reluctant to make a last-minute change we couldn’t adequately test.